Code style update
