+++ src/integrationTest/java/org/radarcns/integration/PhoneStreamTest.java
@@ -55,7 +55,6 @@ import org.radarcns.config.YamlConfigLoader;
 import org.radarcns.kafka.ObservationKey;
 import org.radarcns.mock.MockProducer;
 import org.radarcns.mock.config.BasicMockConfig;
-import org.radarcns.mock.config.MockDataConfig;
 import org.radarcns.monitor.AbstractKafkaMonitor;
 import org.radarcns.monitor.KafkaMonitor;
 import org.radarcns.passive.empatica.EmpaticaE4Acceleration;
@@ -70,7 +69,6 @@ import org.radarcns.producer.KafkaTopicSender;
 import org.radarcns.producer.direct.DirectSender;
 import org.radarcns.schema.registration.KafkaTopics;
 import org.radarcns.schema.registration.SchemaRegistry;
-import org.radarcns.schema.registration.SchemaRegistry.Compatibility;
 import org.radarcns.topic.AvroTopic;
 import org.radarcns.util.RadarSingletonFactory;
 import org.radarcns.util.serde.KafkaAvroSerializer;
