+++ src/main/java/org/radarcns/config/ConfigRadar.java
@@ -18,14 +18,13 @@ package org.radarcns.config;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import org.radarcns.config.RadarPropertyHandler.Priority;
-
 import java.util.Date;
 import java.util.EnumMap;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
+import org.radarcns.config.RadarPropertyHandler.Priority;
 
 /**
  * POJO representing the yml file
