+++ src/main/java/org/radarcns/config/SourceStatisticsMonitorConfig.java
@@ -1,7 +1,6 @@
 package org.radarcns.config;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-
 import java.util.List;
 
 public class SourceStatisticsMonitorConfig {
