+++ src/main/java/org/radarcns/monitor/KafkaMonitorFactory.java
@@ -16,6 +16,14 @@
 
 package org.radarcns.monitor;
 
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.List;
+import java.util.Locale;
+import java.util.stream.Collectors;
 import org.radarcns.config.BatteryMonitorConfig;
 import org.radarcns.config.DisconnectMonitorConfig;
 import org.radarcns.config.MonitorConfig;
@@ -25,15 +33,6 @@ import org.radarcns.config.SourceStatisticsMonitorConfig;
 import org.radarcns.util.EmailSender;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
-import java.util.Locale;
-import java.util.stream.Collectors;
 
 public class KafkaMonitorFactory {
