+++ src/main/java/org/radarcns/stream/empatica/E4StreamMaster.java
@@ -16,13 +16,12 @@
 
 package org.radarcns.stream.empatica;
 
+import java.util.List;
 import org.radarcns.config.RadarPropertyHandler;
 import org.radarcns.stream.StreamGroup;
 import org.radarcns.stream.StreamMaster;
 import org.radarcns.stream.StreamWorker;
 import org.radarcns.util.RadarSingletonFactory;
-
-import java.util.List;
 
 /**
  * Singleton StreamMaster for Empatica E4.
