+++ src/main/java/org/radarcns/stream/phone/PhoneAccelerationStream.java
@@ -7,9 +7,9 @@ import org.radarcns.config.RadarPropertyHandler;
 import org.radarcns.kafka.AggregateKey;
 import org.radarcns.kafka.ObservationKey;
 import org.radarcns.passive.phone.PhoneAcceleration;
+import org.radarcns.stream.KStreamWorker;
 import org.radarcns.stream.StreamDefinition;
 import org.radarcns.stream.StreamMaster;
-import org.radarcns.stream.KStreamWorker;
 import org.radarcns.stream.aggregator.AggregateList;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
