+++ src/test/java/org/radarcns/monitor/DisconnectMonitorTest.java
@@ -16,6 +16,20 @@
 
 package org.radarcns.monitor;
 
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.hasEntry;
+import static org.hamcrest.Matchers.hasKey;
+import static org.hamcrest.Matchers.is;
+import static org.junit.Assert.assertEquals;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+
+import java.io.File;
+import java.util.Collections;
+import java.util.Map;
+import javax.mail.MessagingException;
 import org.apache.avro.Schema;
 import org.apache.avro.Schema.Parser;
 import org.apache.avro.generic.GenericData.Record;
@@ -34,23 +48,7 @@ import org.radarcns.kafka.ObservationKey;
 import org.radarcns.monitor.DisconnectMonitor.DisconnectMonitorState;
 import org.radarcns.monitor.DisconnectMonitor.MissingRecordsReport;
 import org.radarcns.util.EmailSender;
-import org.radarcns.util.PersistentStateStore;
-
-import javax.mail.MessagingException;
-import java.io.File;
-import java.util.Collections;
-import java.util.Map;
-
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.hasEntry;
-import static org.hamcrest.Matchers.hasKey;
-import static org.hamcrest.Matchers.is;
-import static org.junit.Assert.assertEquals;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.radarcns.util.PersistentStateStore.measurementKeyToString;
+import org.radarcns.util.YamlPersistentStateStore;
 
 public class DisconnectMonitorTest {
     @Rule
@@ -154,27 +152,27 @@ public class DisconnectMonitorTest {
     @Test
     public void retrieveState() throws Exception {
         File base = folder.newFolder();
-        PersistentStateStore stateStore = new PersistentStateStore(base);
+        YamlPersistentStateStore stateStore = new YamlPersistentStateStore(base);
         DisconnectMonitorState state = new DisconnectMonitorState();
         ObservationKey key1 = new ObservationKey("test", "a", "b");
         ObservationKey key2 = new ObservationKey("test", "b", "c");
         ObservationKey key3 = new ObservationKey("test", "c", "d");
         long now = System.currentTimeMillis();
-        state.getLastSeen().put(measurementKeyToString(key1), now);
-        state.getLastSeen().put(measurementKeyToString(key2), now + 1L);
-        state.getReportedMissing().put(measurementKeyToString(key3), new MissingRecordsReport
+        state.getLastSeen().put(stateStore.keyToString(key1), now);
+        state.getLastSeen().put(stateStore.keyToString(key2), now + 1L);
+        state.getReportedMissing().put(stateStore.keyToString(key3), new MissingRecordsReport
                 (now -60L, now + 2L, 0));
         stateStore.storeState("one", "two", state);
 
-        PersistentStateStore stateStore2 = new PersistentStateStore(base);
+        YamlPersistentStateStore stateStore2 = new YamlPersistentStateStore(base);
         DisconnectMonitorState state2 = stateStore2.retrieveState("one", "two", new DisconnectMonitorState());
         Map<String, Long> lastSeen = state2.getLastSeen();
         assertThat(lastSeen.size(), is(2));
-        assertThat(lastSeen, hasEntry(measurementKeyToString(key1), now));
-        assertThat(lastSeen, hasEntry(measurementKeyToString(key2), now + 1L));
+        assertThat(lastSeen, hasEntry(stateStore.keyToString(key1), now));
+        assertThat(lastSeen, hasEntry(stateStore.keyToString(key2), now + 1L));
         Map<String, MissingRecordsReport> reported = state2.getReportedMissing();
         assertThat(reported.size(), is(1));
-        assertThat(reported, hasKey(measurementKeyToString(key3)));
+        assertThat(reported, hasKey(stateStore.keyToString(key3)));
 
     }
 }
