+++ src/test/java/org/radarcns/util/PersistentStateStoreTest.java
@@ -20,7 +20,6 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasEntry;
 import static org.hamcrest.core.Is.is;
-import static org.radarcns.util.PersistentStateStore.measurementKeyToString;
 
 import java.io.File;
 import java.nio.file.Files;
@@ -38,10 +37,10 @@ public class PersistentStateStoreTest {
     @Test
     public void retrieveState() throws Exception {
         File base = folder.newFolder();
-        PersistentStateStore stateStore = new PersistentStateStore(base);
+        YamlPersistentStateStore stateStore = new YamlPersistentStateStore(base);
         BatteryLevelState state = new BatteryLevelState();
         ObservationKey key1 = new ObservationKey("test", "a", "b");
-        state.updateLevel(key1, 0.1f);
+        state.updateLevel(stateStore.keyToString(key1), 0.1f);
         stateStore.storeState("one", "two", state);
 
         File outputFile = new File(base, "one_two.yml");
@@ -49,9 +48,9 @@ public class PersistentStateStoreTest {
         String rawFile = new String(Files.readAllBytes(outputFile.toPath()));
         assertThat(rawFile, equalTo("---\nlevels:\n  test#a#b: 0.1\n"));
 
-        PersistentStateStore stateStore2 = new PersistentStateStore(base);
+        YamlPersistentStateStore stateStore2 = new YamlPersistentStateStore(base);
         BatteryLevelState state2 = stateStore2.retrieveState("one", "two", new BatteryLevelState());
         Map<String, Float> values = state2.getLevels();
-        assertThat(values, hasEntry(measurementKeyToString(key1), 0.1f));
+        assertThat(values, hasEntry(stateStore.keyToString(key1), 0.1f));
     }
 }
